test(task-gui-unit-target): 新增无窗口 gui_unit_test target 并迁入 defaults_diff 15 用例
src/gui/ 的纯逻辑测试此前只有 gui_test 一个家，而 gui_test 从不在 CI 执行
（ci.yml build-only，runner 无 display server）。新增第二个 gtest target，
链 lumice_gui_obj 但不开窗口 / 不建 GL 上下文 / 不带 ImGui 测试引擎。

- test/CMakeLists.txt: gui_unit_test（BUILD_GUI 门内）+ add_test(LumiceGuiUnitTest)
  + LABEL "unit-correctness" —— 沿用既有 LABEL 使 ctest -L 选择器零改动即选中
- user_defaults_test_env.hpp 迁到 test/support/（两个 target 共享，layer-agnostic），
  去掉为一个 namespace alias 而生的 test_gui_shared.hpp 依赖
- test_gui_defaults_diff.cpp 的 15 个用例按固定映射迁到
  test/unit-correctness/gui/test_defaults_diff.cpp，原文件删除（是移动不是复制）

核验：断言 92/19/2/9 逐条对应（122）；gui_test 侧 IM_REGISTER_TEST 575→560（-15）；
15/15 绿 8ms，--gtest_shuffle --gtest_repeat=3 稳定；gui_test 剩余消费者 72/72 绿。
