+++ test/gui/CMakeLists.txt
@@ -19,7 +19,6 @@ add_executable(gui_test
     functional/test_gui_state_reconcile.cpp
     functional/test_gui_sampling_density_stats.cpp
     functional/test_gui_user_defaults.cpp
-    functional/test_gui_defaults_diff.cpp
     functional/test_gui_defaults_panel.cpp
     functional/test_project_world_dir.cpp
     functional/test_render_handedness_guard.cpp
@@ -43,8 +42,12 @@ add_executable(gui_test
     ${IMGUI_TE_DIR}/imgui_te_perftool.cpp
     ${IMGUI_TE_DIR}/imgui_capture_tool.cpp)
 
+# ${PROJ_TEST_DIR} (the test/ root) is on the list so gui_test can reach the target-agnostic
+# helpers under test/support/ by their "support/..." path — the same spelling gui_unit_test uses,
+# which keeps one include line valid in both targets.
 target_include_directories(gui_test PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}
+    ${PROJ_TEST_DIR}
     ${IMGUI_TE_DIR})
 
 # `lumice_obj` (objects), not `lumice`: the white-box tests reference internal core
